Fix SDO transfer exceptions and add new tests
When reading or writing to a subindex > 1, an exception would be thrown,
as "subIndex" is not defined, and should instead be "transfer.subIndex".
Commit also includes new tests that verify these now work correctly.
Also changes the SDO timeout test to verify it is the right exception.
Lastly fixes EDS.load failing to create a raw for an object because
DataType was not parsed as an int.